requirements.txt
streamlit
openai-whisper
torch
reportlab
audio-recorder-streamlit
ollama # optional for local LLM
pydub
python-docx
rapidfuzz
